Improve performance by not loading stuff for old explore page on initial group load
However, make explore page available again (for farms)
