+++ apps/web/src/routes/LandingPage/LandingPage.js
@@ -1,10 +1,11 @@
 import React, { useEffect, useMemo } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
-import { useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import Loading from 'components/Loading'
 import Widget from 'components/Widget'
 import { useViewHeader } from 'contexts/ViewHeaderContext'
+import useRouteParams from 'hooks/useRouteParams'
+import fetchGroupForLandingPage from 'store/actions/fetchGroupForLandingPage'
 import fetchPosts from 'store/actions/fetchPosts'
 import presentGroup from 'store/presenters/presentGroup'
 import presentPost from 'store/presenters/presentPost'
@@ -18,7 +19,7 @@ import { RESP_ADMINISTRATION, RESP_MANAGE_CONTENT } from 'store/constants'
 
 const LandingPage = () => {
   const dispatch = useDispatch()
-  const params = useParams()
+  const params = useRouteParams()
   const { t } = useTranslation()
 
   const groupSlug = params.groupSlug
@@ -40,7 +41,8 @@ const LandingPage = () => {
 
   useEffect(() => {
     dispatch(fetchPosts(fetchPostsParam))
-  }, [dispatch, fetchPostsParam])
+    dispatch(fetchGroupForLandingPage({ slug: groupSlug }))
+  }, [fetchPostsParam, groupSlug])
 
   const { setHeaderDetails } = useViewHeader()
   useEffect(() => {
@@ -51,7 +53,7 @@ const LandingPage = () => {
     })
   }, [])
 
-  if (!group) return <Loading />
+  if (!group || widgets.length === 0) return <Loading />
 
   return (
     <div>
