+++ apps/web/src/store/constants.js
@@ -56,6 +56,7 @@ export const FETCH_CONTEXT_WIDGETS = 'FETCH_CONTEXT_WIDGETS'
 export const FETCH_CHILD_COMMENTS = 'FETCH_CHILD_COMMENTS'
 export const FETCH_GROUP_DETAILS = 'FETCH_GROUP_DETAILS'
 export const FETCH_GROUP_DETAILS_PENDING = 'FETCH_GROUP_DETAILS_PENDING'
+export const FETCH_GROUP_FOR_LANDING_PAGE = 'FETCH_GROUP_FOR_LANDING_PAGE'
 export const FETCH_GROUPS = 'FETCH_GROUPS'
 export const FETCH_GROUP_TOPIC = 'FETCH_GROUP_TOPIC'
 export const FETCH_DEFAULT_TOPICS = 'FETCH_DEFAULT_TOPICS'
